Initial Commit
In 2018 NFL season, one of the best running backs, Le'Veon Bell from Pittsburg Steelers decided to sit out the entire season, giving up $14.544 million dollars due to contract dispute. Many argue that running backs are replaceable and they should not get paid as much as Quarterbacks or Wide Receivers. I've compared the data with the backup running back (James Conner) with Le'Veon Bell to see if this was true.